requirements.txt
asgiref==3.2.3
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
configobj==5.0.6
cryptography==2.9
Django==3.0.3
django-extensions==3.0.3
django-werkzeug-debugger-runserver==0.3.1
idna==2.8
libvirt-python==6.0.0
paramiko==2.7.1
pycparser==2.20
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
Werkzeug==1.0.1
docker==4.3.1